Split up functions into separate files
